refactor!: change all int -> long

+++ src/main/java/dev/razafindratelo/set/Z.java
@@ -3,11 +3,11 @@ package dev.razafindratelo.set;
 import dev.razafindratelo.utils.Euclidian;
 
 /**
- *  Z stands for the set of negative and positive integer as in Mathematics
+ *  Z stands for the set of negative and positive Integer as in Mathematics
  */
 public class Z implements R {
 
-    public static int gcd(int a, int b) {
+    public static long gcd(long a, long b) {
         var euclidian =  Euclidian.division(a, b);
         if (euclidian.get("r") != 0) {
             return gcd(b, euclidian.get("r"));
@@ -16,7 +16,7 @@ public class Z implements R {
         }
     }
 
-    public static int gcm(int a, int b) {
+    public static long gcm(long a, long b) {
         return (a * b) / gcd(a, b);
     }
 }
